refactor command registry to use model for params
