+++ src/commands/models/command_mapping.py
@@ -1,11 +1,11 @@
 from typing import Callable, List, TypedDict
 
 from discord import Message
-from discord.app_commands import Parameter
+from commands.models.command_param import CommandParam
 
 class CommandEntry(TypedDict):
     function: Callable[[dict], Message]
     description: str
-    params: List[Parameter]
+    params: List[CommandParam]
 
 CommandMapping = dict[str, CommandEntry]
